Add GUA to geomagio/Metadata.py
GUA is transitioning to an ObsRIO data logger, which requires a
metadata entry in Metadata.py. We do not know actual calibration
values for the GUA Narod mag, and cannot retrieve them currently
due to ongoing communication issues, so we'll use Narod default
calibration parameters.

This commit applies strictly to GUA. Additional commits are
forthcoming that relate to the test FGE+ObsRIO system that will
also be deployed soon.